Remove get static props from token page
