+++ src/app/t/[network]/[contractAddress]/page.tsx
@@ -1,23 +1,17 @@
-'use client'
-
-import ContractDefinedSpace from "@/common/components/pages/ContractDefinedSpace";
-import SpaceNotFound from "@/common/components/pages/SpaceNotFound";
 import { OwnerType } from "@/common/data/api/etherscan";
 import { loadContractData } from "@/common/data/loaders/contractPagePropsLoader";
-import { useAppStore } from "@/common/data/stores/app";
-import { generateContractMetadataHtml } from "@/common/lib/utils/generateContractMetadataHtml";
 import { NextPageWithLayout } from "@/pages/_app";
-import { isArray, isNil } from "lodash";
-import { GetServerSideProps, GetServerSidePropsContext } from "next";
-import Head from "next/head";
-import React, { useEffect } from "react";
+import React from "react";
 import {
   MasterToken,
   TokenProvider,
-  useToken,
 } from "@/common/providers/TokenProvider";
 import { Address } from "viem";
 import { EtherScanChainName } from "@/constants/etherscanChainIds";
+import { useParams } from 'next/navigation';
+import { ContractPrimarySpaceContent } from '../ContractPrimarySpaceContent';
+import { Metadata } from 'next/types';
+import { generateContractMetadataHtml } from '@/common/lib/utils/generateContractMetadataHtml';
 
 export interface ContractSpacePageProps {
   spaceId: string | null;
@@ -31,28 +25,25 @@ export interface ContractSpacePageProps {
   network: EtherScanChainName;
 }
 
-export const getServerSideProps = (async ({
-  params,
-}: GetServerSidePropsContext) => {
-  const contractData = await loadContractData(params);
-  const network = params?.network as EtherScanChainName;
-  return {
-    props: {
-      ...contractData.props,
-      network,
-    },
-  };
-}) satisfies GetServerSideProps<ContractSpacePageProps>;
+export async function generateMetadata({
+  params: { contractAddress, network },
+}): Promise<Metadata> {
+  const metadata = generateContractMetadataHtml(contractAddress);
+  return metadata;
+}
 
-export const ContractPrimarySpace: NextPageWithLayout = ({
-  spaceId,
-  tabName,
-  ownerId,
-  ownerIdType,
-  contractAddress,
-  owningIdentities,
-  network,
-}: ContractSpacePageProps) => {
+export const ContractPrimarySpace: NextPageWithLayout = async () => {
+  const params = useParams();
+  const {props: {
+    spaceId,
+    tabName,
+    ownerId,
+    ownerIdType,
+    contractAddress,
+    owningIdentities,
+  }} = await loadContractData(params || {});
+  const network = params?.network as EtherScanChainName;
+
   return (
     <TokenProvider
       contractAddress={contractAddress as Address}
@@ -69,57 +60,6 @@ export const ContractPrimarySpace: NextPageWithLayout = ({
       />
     </TokenProvider>
   );
-};
-
-export const ContractPrimarySpaceContent: React.FC<ContractSpacePageProps> = ({
-  spaceId,
-  tabName,
-  ownerId,
-  ownerIdType,
-  contractAddress,
-  owningIdentities,
-  network,
-}) => {
-  const { loadEditableSpaces, addContractEditableSpaces } = useAppStore(
-    (state) => ({
-      loadEditableSpaces: state.space.loadEditableSpaces,
-      addContractEditableSpaces: state.space.addContractEditableSpaces,
-    }),
-  );
-
-  const { tokenData } = useToken();
-
-  useEffect(() => {
-    addContractEditableSpaces(spaceId, owningIdentities);
-  }, [spaceId]);
-
-  useEffect(() => {
-    loadEditableSpaces();
-  }, []);
-
-  if (!isNil(ownerId) && !isNil(contractAddress)) {
-    if (
-      (isNil(spaceId) &&
-        (tabName?.toLocaleLowerCase() === "profile" || tabName === null)) ||
-      !isNil(spaceId)
-    )
-      return (
-        <>
-          <Head>
-            {generateContractMetadataHtml(contractAddress, tokenData)}
-          </Head>
-          <ContractDefinedSpace
-            ownerId={ownerId}
-            ownerIdType={ownerIdType}
-            spaceId={spaceId}
-            tabName={isArray(tabName) ? tabName[0] : tabName ?? "Profile"}
-            contractAddress={contractAddress}
-          />
-        </>
-      );
-  }
-
-  return <SpaceNotFound />;
 };
 
 export default ContractPrimarySpace;
