+++ src/common/lib/utils/generateContractMetadataHtml.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { MasterToken } from "@/common/providers/TokenProvider";
+import { Metadata } from 'next/types';
 
 export type UserMetadata = {
   username?: string;
@@ -11,7 +12,7 @@ export type UserMetadata = {
 export const generateContractMetadataHtml = (
   contractAddress?: string | null,
   tokenData?: MasterToken | null,
-) => {
+): Metadata => {
   const spaceUrl = `https://nounspace.com/t/${tokenData?.network}/${contractAddress}`;
   const priceInfo = tokenData?.geckoData?.price_usd
     ? ` - $${Number(tokenData.geckoData?.price_usd)} USD`
@@ -22,14 +23,18 @@ export const generateContractMetadataHtml = (
 
   const title = `${symbol}${priceInfo}`;
 
-  return (
-    <>
-      <title>{title}</title>
-      <meta property="og:title" content={`${title}`} />
-      <meta name="twitter:title" content={`${title}`} />
-      <meta property="twitter:domain" content="https://nounspace.com/" />
-      <meta property="og:url" content={spaceUrl} />
-      <meta property="twitter:url" content={spaceUrl} />
-    </>
-  );
+  const metadata = {
+    title,
+    openGraph: {
+      title,
+      url: spaceUrl,
+    },
+    twitter: {
+      title,
+      domain: spaceUrl,
+      url: spaceUrl,
+    },
+  };
+
+  return metadata;
 };
